global effects provider 1/5: make effectsprovider global: move `performEffect` to the scenario
