+++ src/effects/effects_base/internal/effectsprovider.h
@@ -7,7 +7,6 @@
 #include "framework/global/modularity/ioc.h"
 #include "framework/interactive/iinteractive.h"
 
-#include "context/iglobalcontext.h"
 #include "playback/iplayback.h"
 
 #include "audioplugins/iknownaudiopluginsregister.h"
@@ -19,6 +18,7 @@
 
 class EffectBase;
 class EffectSettingsAccess;
+class TrackList;
 
 namespace BasicUI {
 class ProgressDialog;
@@ -31,7 +31,6 @@ class EffectsProvider : public IEffectsProvider, public muse::async::Asyncable, 
     muse::GlobalInject<muse::audioplugins::IKnownAudioPluginsRegister> knownPluginsRegister;
     muse::GlobalInject<IEffectLoadersRegister> effectLoadersRegister;
 
-    muse::ContextInject<au::context::IGlobalContext> globalContext{ this };
     muse::ContextInject<muse::IInteractive> interactive{ this };
     muse::ContextInject<playback::IPlayback> playback{ this };
     muse::ContextInject<IEffectViewLaunchRegister> viewLaunchRegister{ this };
@@ -61,9 +60,6 @@ public:
     void showEffect(const RealtimeEffectStatePtr& state) const override;
     void hideEffect(const RealtimeEffectStatePtr& state) const override;
 
-    muse::Ret performEffect(au3::Au3Project& project, Effect* effect, std::shared_ptr<EffectInstance> effectInstance,
-                            EffectSettings& settings) override;
-
     muse::Ret previewEffect(const EffectId& effectId, EffectSettings& settings) override;
     void stopPreview() override;
 
@@ -71,18 +67,18 @@ private:
     struct EffectContext {
         double t0 = 0.0;
         double t1 = 0.0;
-        std::shared_ptr<TrackList> tracks;
+        std::shared_ptr<::TrackList> tracks;
         BasicUI::ProgressDialog* preparingPreviewProgress = nullptr;
         bool isPreview = false;
     };
 
     struct EffectPreviewState {
         EffectPreviewState(const EffectId& effectId, const EffectContext& originContext,
-                           const std::shared_ptr<TrackList>& previewTracks, bool loopWasActive)
+                           const std::shared_ptr<::TrackList>& previewTracks, bool loopWasActive)
             : effectId(effectId), originContext(originContext), previewTracks(previewTracks), loopWasActive(loopWasActive) {}
         const EffectId effectId;
         const EffectContext originContext;
-        const std::shared_ptr<TrackList> previewTracks;
+        const std::shared_ptr<::TrackList> previewTracks;
         const bool loopWasActive;
     };
